requirements.txt
anyio==3.6.1
async-generator==1.10
attrs==22.1.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi==1.15.1
cryptography==37.0.4
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
outcome==1.2.0
Pillow==9.2.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
rfc3986==1.5.0
selenium==4.3.0
sniffio==1.2.0
sortedcontainers==2.4.0
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
wsproto==1.1.0
